backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
gitpython
langchain
langchain-groq
langgraph
sentence-transformers
chromadb
httpx
langchain-text-splitters
python-multipart
